[JENKINS-16350] fix the logout -> immediate login
When using a secured jenkins without anonymous read on the main page the
redirect at the end of the login would case Jenkins to redirect to the
security realms login page.

This does not log you out of github but at least it means your JENKINS
cookie can not be snarfed by anyone for evil purposes.